requirements.txt
# Azure SDK
azure-storage-blob==12.19.0
azure-identity==1.15.0

# Data Processing
pandas==2.1.4
numpy==1.26.2
pyarrow==14.0.1

# Data Quality
great-expectations==0.18.8

# Machine Learning
scikit-learn==1.3.2
xgboost==2.0.3
lightgbm==4.1.0
imbalanced-learn==0.11.0

# MLOps
mlflow==2.9.2

# Model Explainability
shap==0.44.0
evidently==0.4.12

# Visualization
matplotlib==3.8.2
seaborn==0.13.1

# Utilities
python-dotenv==1.0.0